move iam package to top-level
Also setup bintray
