+++ cf/src/main/scala/typeformation/iam/Condition.scala
@@ -1,4 +1,4 @@
-package typeformation.cf.iam
+package typeformation.iam
 
 import io.circe.Encoder
