+++ cf/src/main/scala/typeformation/iam/Invertible.scala
@@ -1,4 +1,4 @@
-package typeformation.cf.iam
+package typeformation.iam
 
 import typeformation.cf.Arn
