+++ cf/src/main/scala/typeformation/iam/Principal.scala
@@ -1,4 +1,4 @@
-package typeformation.cf.iam
+package typeformation.iam
 
 import typeformation.cf.Arn
 import typeformation.cf.CfExp
