+++ cf/src/main/scala/typeformation/iam/syntax/package.scala
@@ -1,3 +1,3 @@
-package typeformation.cf.iam
+package typeformation.iam
 
 package object syntax extends Conversions with Conditions
