+++ project/plugins.sbt
@@ -1,1 +1,3 @@
 logLevel := Level.Warn
+
+addSbtPlugin("org.foundweekends" % "sbt-bintray" % "0.5.1")
